Added Connection.insert_batch(ns, {doc1, doc2,...})
Batching is more efficient for when there are many documents to insert at one time.
The protocol overhead is the same, but it is less kernel/userspace
transitions for the networking.

There may be a message size limit.  This is not checked, but simply delegates handling to the C++ API.